Remove ArrayAccessData concern.
